refactor: consolidate date utilities into shared file
